requirements.txt

# requirements.txt
streamlit
pdfplumber
transformers
pandas
mysql-connector-python
pdfplumber
openpyxl